Fix broken loading of local modules under Perl 5.26
Perl 5.26 breaks the implication that:

  use inc::Module::Install;
  use t::lib::Debugger;

Will load:

  ./inc/Module/Install.pm
  ./t/lib/Debugger.pm

Respectively, due to '.' ceasing to be in @INC

This fixes:
- Makefile.PL by re-inserting the '.' ( which is the only
  thing that works, due to Module::Install shenanigans ).
- Tests by replacing "use" statements with equivalent "require"
statements.

Some of the existing code ( eg: use_ok ) was already spurious and not
very smart, because calling ->import while being outside of BEGIN { }
has limited usefullness.

But this was left with the semantically equivalent code that retains
the loading of the relative path.

There are strategies that would be "nicer" than what I've done,
but they all wind up with you wanting to rename "Debugger.pm" to
something else, because:

  use lib "t/lib";
  use Debugger;

Is going to give people a much different impression from either

  use t::lib::Debugger

Or

  BEGIN {
    require "./t/lib/Debugger.pm";
    t::lib::Debugger->import();
  }

This closes https://github.com/PadreIDE/Debug-Client/issues/6